item refactoring
Took 1 hour 44 minutes

+++ amoginarium/shared/_shared_memory.py
@@ -61,7 +61,7 @@ class base_controller_t(ctypes.Structure):
 class item_slot_t(ctypes.Structure):
     _pack_ = 1
     _fields_ = [
-        ("item_id", ctypes.c_uint8),
+        ("item_id", ctypes.c_uint16),
         ("count", ctypes.c_uint8)
     ]
